pyproject.toml
[tool.poetry]
name = "kg-integration"
version = "1.2.2"
description = "Service for performing KG requests."
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.95"
FastAPI-SQLAlchemy = "0.2.1"
pydantic = "1.10.9"
uvicorn = { extras = ["standard"], version = "^0.23" }
pilot-platform-common = "0.4.1"
psycopg2 = "2.9.7"
asyncpg = "0.28.0"
backoff = "2.2.1"

[tool.poetry.dev-dependencies]
SQLAlchemy-Utils = "0.38.2"
alembic = "1.11.2"
pytest-asyncio = "0.21.1"
testcontainers = "3.7.1"
pytest-httpx = "^0.21"

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
